Move useful types to :test-lib
