+++ dialogue-serde/build.gradle
@@ -8,6 +8,7 @@ dependencies {
     compile 'com.palantir.safe-logging:preconditions'
     compile 'org.slf4j:slf4j-api'
 
+    testCompile project(':dialogue-client-test-lib')
     testCompile 'junit:junit'
     testCompile 'org.assertj:assertj-core'
     testCompile 'org.immutables:value::annotations'
