+++ dialogue-serde/src/test/java/com/palantir/conjure/java/dialogue/serde/DefaultClientsTest.java
@@ -30,11 +30,13 @@ import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.SettableFuture;
 import com.palantir.dialogue.BodySerDe;
 import com.palantir.dialogue.Channel;
+import com.palantir.dialogue.CloseRecordingInputStream;
 import com.palantir.dialogue.Deserializer;
 import com.palantir.dialogue.Endpoint;
 import com.palantir.dialogue.Request;
 import com.palantir.dialogue.RequestBody;
 import com.palantir.dialogue.Response;
+import com.palantir.dialogue.TestResponse;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Optional;
